I has been finished  the sevice layer

+++ src/main/java/com/kainom/shop/dto/ItemDTO.java
@@ -1,5 +1,5 @@
 package com.kainom.shop.dto;
 
 
-public record ItemDTO(String productIdentifier,Float price) {
+public record ItemDTO(String productIdentifier,Double price) {
 }
